# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.AllergenResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.CreateTakeoutOrderRequest:
    properties:
      notes:
        type: string
    type: object
  github_com_AboloreDev_geritcht-restaurant_internals_dto.DietaryTagResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.MenuCategoryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      display_order:
        type: integer
      id:
        type: integer
      image_url:
        type: string
      is_active:
        type: boolean
      menu_items:
        items:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.MenuResponse'
        type: array
      name:
        type: string
    type: object
  github_com_AboloreDev_geritcht-restaurant_internals_dto.MenuImageResponse:
    properties:
      alt_text:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      url:
        type: string
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.MenuResponse:
    properties:
      allergens:
        items:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.AllergenResponse'
        type: array
      category:
        $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.MenuCategoryResponse'
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      dietary_tags:
        items:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.DietaryTagResponse'
        type: array
      display_order:
        type: integer
      id:
        type: integer
      image_url:
        type: string
      images:
        items:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.MenuImageResponse'
        type: array
      is_available:
        type: boolean
      name:
        type: string
      prep_time_minutes:
        type: integer
      price:
        type: number
      spice_level:
        type: integer
      updated_at:
        type: string
    type: object
  github_com_AboloreDev_geritcht-restaurant_internals_dto.OrderItemResponse:
    properties:
      id:
        type: integer
      menu_item:
        $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.MenuResponse'
      menu_item_id:
        type: integer
      price:
        type: number
      quantity:
        type: integer
      special_instructions:
        type: string
      subtotal:
        type: number
    type: object
  github_com_AboloreDev_geritcht-restaurant_internals_dto.OrderResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      notes:
        type: string
      order_items:
        items:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.OrderItemResponse'
        type: array
      payment:
        $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.PaymentResponse'
      payment_status:
        type: string
      reservation_id:
        type: integer
      status:
        type: string
      table_id:
        type: integer
      total_amount:
        type: number
      type:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.UserResponse'
      user_id:
        type: integer
    type: object
  github_com_AboloreDev_geritcht-restaurant_internals_dto.PaymentResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      currency:
        type: string
      failure_reason:
        type: string
      id:
        type: integer
      order_id:
        type: integer
      paid_at:
        type: string
      provider:
        type: string
      provider_reference:
        type: string
      reference:
        type: string
      status:
        type: string
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      first_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_name:
        type: string
      phone_number:
        type: string
      role:
        type: string
    type: object
...
  github_com_AboloreDev_geritcht-restaurant_internals_utils.PaginatedMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_AboloreDev_geritcht-restaurant_internals_utils.PaginatedResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.PaginatedMeta'
      status:
        type: boolean
    type: object
  github_com_AboloreDev_geritcht-restaurant_internals_utils.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      status:
        type: boolean
    type: object
host: https://geritcht-restaurant.onrender.com
info:
  contact:
    email: alabifathiu@yahoo.com
    name: Alabi Fathiu
    url: http://linkedin.com/in/fathiu-alabi
  description: A modern Restaurant API built with Go, Gin, and GORM
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Geritcht Restaurant API
  version: "1.0"
paths:
...
  /orders/all:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all orders placed by the authenticated
        user.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Orders retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.OrderResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Get my takeout orders
      tags:
      - Orders
  /orders/search:
    get:
      consumes:
      - application/json
      description: Search orders by name with pagination
      parameters:
      - description: Order name
        in: query
        name: query
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Orders retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.OrderResponse'
                  type: array
              type: object
        "400":
          description: Invalid search parameters
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: No orders found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Search Orders
      tags:
      - Orders
  /orders/takeout:
    post:
      consumes:
      - application/json
      description: Create a new takeout order from the authenticated user's cart.
      parameters:
      - description: Takeout order details
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.CreateTakeoutOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Order created successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.OrderResponse'
              type: object
        "400":
          description: Invalid request data, cart is empty, or menu item unavailable
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: Cart not found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Create a takeout order
      tags:
      - Orders
  /orders/takeout/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve the details of a specific takeout order placed by the
        authenticated user.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.OrderResponse'
              type: object
        "400":
          description: Invalid order ID
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Get a takeout order
      tags:
      - Orders
  /orders/takeout/{id}/cancel:
    patch:
      consumes:
      - application/json
      description: Cancel a takeout order placed by the authenticated user if it's
        eligible for cancellation.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Order cancelled successfully
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "400":
          description: Invalid order ID or order cannot be cancelled
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Cancel a takeout order
      tags:
      - Orders
  /orders/takeout/all:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all takeout orders placed by the authenticated
        user.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Orders retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_dto.OrderResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_AboloreDev_geritcht-restaurant_internals_utils.Response'
      security:
      - BearerAuth: []
      summary: Get my takeout orders
      tags:
      - Orders
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"